refactor(mobile): logically split code
