+++ apps/mobile/src/services/analytics/index.ts
@@ -1,3 +1,5 @@
 export * from './events'
 export * from './types'
+export * from './constants'
+export * from './utils'
 export * from './firebaseAnalytics'
