+++ apps/mobile/src/store/middleware/analytics/strategies/TransactionConfirmationStrategy.ts
@@ -2,7 +2,7 @@ import { RootState } from '@/src/store'
 import { MiddlewareAPI, Dispatch } from 'redux'
 import { Strategy, ActionWithPayload } from '@/src/store/utils/strategy/Strategy'
 import { trackEvent } from '@/src/services/analytics'
-import { getTransactionAnalyticsLabel } from '@/src/services/analytics/types'
+import { getTransactionAnalyticsLabel } from '@/src/services/analytics/utils'
 import { createTxConfirmEvent } from '@/src/services/analytics/events/transactions'
 import type { Transaction } from '@safe-global/store/gateway/AUTO_GENERATED/transactions'
